jakarta migration

+++ spring-reactive-jdbc/src/main/java/de/freese/spring/reactive/EmployeeService.java
@@ -1,13 +1,12 @@
 package de.freese.spring.reactive;
 
-import javax.annotation.Resource;
-
-import org.springframework.stereotype.Service;
-import org.springframework.transaction.annotation.Transactional;
+import jakarta.annotation.Resource;
 
 import de.freese.spring.reactive.model.Department;
 import de.freese.spring.reactive.model.Employee;
 import de.freese.spring.reactive.repository.EmployeeRepository;
+import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 import reactor.core.publisher.Flux;
 import reactor.core.publisher.Mono;
 
@@ -17,56 +16,31 @@ import reactor.core.publisher.Mono;
 @Service
 public class EmployeeService
 {
-    /**
-     *
-     */
     @Resource
     private EmployeeRepository repository;
 
-    /**
-     * @param newEmployee {@link Employee}
-     *
-     * @return {@link Mono}
-     */
     @Transactional
     public Mono<Employee> createNewEmployee(final Employee newEmployee)
     {
         return this.repository.createNewEmployee(newEmployee);
     }
 
-    /**
-     * @param id long
-     *
-     * @return {@link Mono}
-     */
     @Transactional
-    public Mono<Integer> deleteEmployee(final long id)
+    public Mono<Long> deleteEmployee(final long id)
     {
         return this.repository.deleteEmployee(id);
     }
 
-    /**
-     * @return {@link Flux}
-     */
     public Flux<Department> getAllDepartments()
     {
         return this.repository.getAllDepartments();
     }
 
-    /**
-     * @return {@link Flux}
-     */
     public Flux<Employee> getAllEmployees()
     {
         return this.repository.getAllEmployees();
     }
 
-    /**
-     * @param lastName String
-     * @param firstName String
-     *
-     * @return {@link Mono}
-     */
     public Mono<Employee> getEmployee(final String lastName, final String firstName)
     {
         return this.repository.getEmployee(lastName, firstName);
